% Generated by roxygen2 (4.0.1): do not edit by hand
\name{col2hex}
\alias{col2hex}
\title{Convert colors to hexidecimal codes}
\usage{
col2hex(col, alpha = 255)
}
\arguments{
\item{col}{A vector or matrix of either numeric colors or color names}

\item{alpha}{The alpha level of the output color ranging from [0,255].
  Default is 255 (opaque).}
}
\value{
An object of the same type and dimensions as \code{col} with all entries replaced
with hexidecimal string equivalents
}
\description{
Convert colors to hexidecimal codes
}
\details{
Uses the color's underlying RGB space to generate the hex code.
}
\section{DEPRECATED}{

Use \link[grDevices]{adjustcolor} instead.
}
\seealso{
\link[grDevices]{adjustcolor} performs the same function and has a few more options
}

