% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extrema}
\alias{extrema}
\title{Determine local extrema}
\usage{
extrema(x, y = NULL, ztol = 1e-05, ddytol = NULL, fx = 1000)
}
\arguments{
\item{x}{Either an \link{xy.coords} object or an atomic numeric vector of x-coordinates}

\item{y}{Required if \code{x} is atomic, an atomic numeric vector of y-coordinates
the same length as \code{x}.}

\item{ztol}{Tolerance for "zero" values. Default: 1e-5.}

\item{ytol}{Tolerance for values to be flagged as extrema.
Default (NULL) automatically computes as 10% of the data range.}

\item{fx}{Multiplicative factor by which the number of points in the data is
  expanded (via \link{splinefun}) to refine extrema location detection. Default: 1000.}
}
\value{
An \link{xy.coords} compatible list with the following elements:
\describe{
 \item{x}{x-coordinates of local extrema}
 \item{y}{y-coordinates of local extrema}
 \item{indices}{indices of local extrema in the original data}
 \item{is.maxima}{logical vector the same length as indices where \code{TRUE} indicates a local maximum}
}
}
\description{
Determines the coordinates of local extrema in a series and identifies them as
either maxima or minima.
}
\details{
The data series is first smoothed using adaptive loess and then spline interpolated
between points.  Zero-crossings of the first derivative of spline interpolated data
are extracted as local extrema.  The nearest 10 points (5-pre, 5-post) surrounding
the extrema are used to determine if the extrema is a local minimum or local maximum.
}
\seealso{
\link{xy.coords}, \link{smooth.adaptive.loess}, \link{splinefun}
}

