% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getKeyValue}
\alias{getKeyValue}
\title{Get a key value from a key-value pair list string}
\usage{
getKeyValue(s, key, val.fun = identity, ...)
}
\arguments{
\item{s}{Character vector of key-value pair list strings}

\item{key}{Character vector of length 1 of the key to be extracted}

\item{val.fun}{Function to apply to the extracted values}
}
\value{
A vector the same length as \code{s} of values defined by \code{key} and
and transformed by \code{val.fun}
}
\description{
Get a key value from a key-value pair list string
}
\details{
A wrapper for \link[pangr]{kvp}
}
\seealso{
\link[pangr]{kvp}
}

