% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gg_halign}
\alias{gg_halign}
\title{Horizontally align ggplots}
\usage{
gg_halign(..., heights = NULL, plot = FALSE)
}
\arguments{
\item{...}{\code{ggplot} objects as separate arguments or a single \code{list()}
of ggplot objects to align horizontally}

\item{heights}{Height(s) for each plot specified with \link[grid]{unit} as defined
with \link[grid]{grid.layout}.  Default is \code{NULL}, heights are evenly
distributed among plots.}

\item{plot}{Logical specifying if the result should be plotted.  Default is
   \code{FALSE}}
}
\value{
If \code{plot = FALSE}, an \link[gridExtra]{arrangeGrob} object is returned.
}
\description{
Horizontally align ggplots
}
\seealso{
\link[grid]{grid.layout}, \link[gridExtra]{arrangeGrob}, \link[grid]{unit}
}

