% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gregexcap}
\alias{gregexcap}
\title{Global RegEx processing with named capture}
\usage{
gregexcap(pattern, x, ...)
}
\arguments{
\item{pattern}{Regex pattern with named capturing groups}

\item{x}{Character vector where \code{pattern} is searched for in each element}

\item{...}{Additional arguments passed to \link[base]{gregexpr}.}
}
\value{
A list the same length as \code{x} where each element has named elements
corresponding to the named groups in \code{pattern}.
}
\description{
Global RegEx processing with named capture
}
\details{
A wrapper for \link[base]{gregexpr} that facilitates retrieval of named
capture groups
}
\examples{
> x = '`a` + `[b]` + `[1c]` + `[d] e`'
> z = '`f` + `[g]` + `[1h]` + `[i] j`'
> gregexcap('`(?<tok>.*?)`', x)
[[1]]
[[1]]$tok
[1] "a"     "[b]"   "[1c]"  "[d] e"

> gregexcap('`(?<tok>.*?)`', c(x,z))
[[1]]
[[1]]$tok
[1] "a"     "[b]"   "[1c]"  "[d] e"

[[2]]
[[2]]$tok
[1] "f"      "[g]"    "[1h]"  "[i] j"
}
\seealso{
\link[base]{gregexpr}
}

