% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hline}
\alias{hline}
\title{Create decorative lines for console output}
\usage{
hline(char = "-", width = 80)
}
\arguments{
\item{char}{Character to use to create the line. Default is \code{'-'}}

\item{width}{Width in characters of the line. Default is 80.}
}
\value{
A string with \code{char} replicated \code{width} times that one could use
for console message display or other text based output.
}
\description{
Create decorative lines for console output
}
\details{
A simple wrapper for \code{past(rep(...), collapse='')}.
}

