% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hmcb}
\alias{hmcb}
\title{Heatmap with colorbar}
\usage{
hmcb(M, nlev = 32, colorbar = T, rug = F, zlab = NULL, zticks = 5,
  crp = colorRampPalette(c("white", "red")), ...)
}
\arguments{
\item{M}{Matrix of values to plot}

\item{nlev}{Number of color levels, default: 32.}

\item{colorbar}{Boolean, TRUE to show colorbar}

\item{zlab}{Axis label for colorbar. NULL and NA values suppress label. default: NULL}

\item{zticks}{Number of z-axis tick marks}

\item{crp}{Color ramp function (e.g. as returned by \code{colorRampPalette()})}

\item{...}{Additional arguments passed to \code{image()}}
}
\value{
\code{NULL} is returned invisibly.
}
\description{
Draws a heatmap based on matrix data with an optional associated colorbar
}
\details{
Plots both the matrix heatmap and colorbar (if specified) using \code{image()}.
Placement of each subplot is done using \code{par(fig=...)}.  Thus, this
function is not compatible with \code{layout()} or \code{par(mfrow=..., mfcol=...)}.

Yet to be determined if this function works with \code{split.screen}.
}
\seealso{
\link{image}, \link{colorRampPalette}
}

