% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.between}
\alias{is.between}
\title{Check if a value is within an interval}
\usage{
is.between(x, interval, na.rm = T, endpts = c("both", "lb", "ub", "none"))
}
\arguments{
\item{x}{A numeric vector of values to test}

\item{interval}{A numeric vector with at least 2 unique values that defines
the interval over which the value of \code{x} must reside}

\item{endpts}{A character value that determines how endpoints are treated to
  determine "between-ness":
  \describe{
   \item{both}{(Default) \code{TRUE} if min(interval) <= x <= max(interval)}
   \item{lb}{\code{TRUE} if min(interval) <= x < max(interval)}
   \item{ub}{\code{TRUE} if min(interval) < x <= max(interval)}
   \item{none}{\code{TRUE} if min(interval) < x < max(interval)}
  }}
}
\value{
A logical vector the same size as \code{x} that is \code{TRUE} for elements
of \code{x} that satisfy the conditions defined by \code{interval} and \code{endpts},
and \code{FALSE} otherwise.
}
\description{
Check if a value is within an interval
}

