% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kvp}
\alias{kvp}
\title{Split a key-value pair list string}
\usage{
kvp(s, pair.split = ";", kv.split = ":")
}
\arguments{
\item{s}{Character vector of key-value pair list strings}

\item{pair.split}{Regex pattern that splits individual key-value pairs. Default ";"}

\item{kv.split}{Regex pattern that splits keys and values. Default ":"}
}
\value{
A list the same length as \code{s} whose elements are lists whose names are
extracted keys and elements are extracted values.
}
\description{
Split a key-value pair list string
}
\details{
Essentially a recursive wrapper for \code{strsplit} that is \code{lapply}'d over \code{s}
}
\seealso{
\link{strsplit}, \link{lapply}.
In addtion, \link[pangr]{getKeyValue} is a wrapper that reduces parsing and
extracting a single key-value to a single function call.
}

