% Generated by roxygen2 (4.0.1): do not edit by hand
\name{multiplot}
\alias{multiplot}
\title{Layout multiple ggplots}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects provided as separate arguments}

\item{plotlist}{A list of ggplot objects}

\item{cols}{Number of columns the layout should have}

\item{layout}{A layout matrix. If present \code{cols} is ignored}
}
\description{
Layout multiple ggplots
}
\details{
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}
\section{Credits}{

This function was copied verbatim from the cookbook for R:
\link{http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}

Documentation was modified to conform to \link{roxygen2} tagging standards
}

