% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summarySE}
\alias{summarySE}
\title{Summarize data}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{data}{Input \code{data.frame}}

\item{measurevar}{Column with variable to be summarized}

\item{groupvars}{Vector of column names that contain grouping variables}

\item{na.rm}{Boolean flag to ignore \code{NA} values}

\item{conf.interval}{Percent range of the inferred confidence interval (default: 95%)}
}
\description{
Determine count, mean, standard deviation, standard error of the mean, and confidence interval (default 95%).
}
\section{Credits}{

This function was copied verbatim from the cookbook for R:
\link{http://www.cookbook-r.com/Graphs/Plotting_means_and_error_bars_(ggplot2)/}

Documentation was modified to conform to \link{roxygen2} tagging standards
}

